refactor(frontend): split TrendsForecastsPage into trends-forecasts/ module
