+++ frontend/src/App.tsx
@@ -31,7 +31,7 @@ const pageImports = {
   SpendingAnalysisPage: () => import('@/pages/SpendingAnalysisPage'),
   IncomeAnalysisPage: () => import('@/pages/IncomeAnalysisPage'),
   IncomeExpenseFlowPage: () => import('@/pages/IncomeExpenseFlowPage'),
-  TrendsForecastsPage: () => import('@/pages/TrendsForecastsPage'),
+  TrendsForecastsPage: () => import('@/pages/trends-forecasts/TrendsForecastsPage'),
   ComparisonPage: () => import('@/pages/ComparisonPage'),
   BudgetPage: () => import('@/pages/BudgetPage'),
   YearInReviewPage: () => import('@/pages/year-in-review/YearInReviewPage'),
