--- frontend/src/pages/TrendsForecastsPage.tsx
@@ -1,775 +1,0 @@
-import { motion } from 'framer-motion'
-import { rawColors } from '@/constants/colors'
-import { TrendingUp, TrendingDown, Minus, Wallet, PiggyBank, CreditCard, LineChart, ArrowUpRight, ArrowDownRight } from 'lucide-react'
-import { useTrends } from '@/hooks/useAnalytics'
-import { useChartDimensions } from '@/hooks/useChartDimensions'
-import { Area, AreaChart, XAxis, YAxis, CartesianGrid, Tooltip, Line, ReferenceLine } from 'recharts'
-import { getDateKey } from '@/lib/dateUtils'
-import { useState, useMemo } from 'react'
-import { formatCurrency, formatCurrencyShort, formatPercent, percentChange } from '@/lib/formatters'
-import { chartTooltipProps, PageHeader, ChartContainer, GRID_DEFAULTS, xAxisDefaults, yAxisDefaults, areaGradient, areaGradientUrl, shouldAnimate, ACTIVE_DOT } from '@/components/ui'
-import { CashFlowForecast } from '@/components/analytics'
-import EmptyState from '@/components/shared/EmptyState'
-import ChartEmptyState from '@/components/shared/ChartEmptyState'
-import AnalyticsTimeFilter from '@/components/shared/AnalyticsTimeFilter'
-import { useTransactions } from '@/hooks/api/useTransactions'
-import { usePreferences } from '@/hooks/api/usePreferences'
-import { useAnalyticsTimeFilter } from '@/hooks/useAnalyticsTimeFilter'
-
-type TrendDirection = 'up' | 'down' | 'stable'
-
-interface TrendMetrics {
-  current: number
-  previous: number
-  change: number
-  changePercent: number
-  direction: TrendDirection
-  average: number
-  highest: number
-  lowest: number
-}
-
-function getDirectionIcon(direction: TrendDirection): React.ReactElement {
-  if (direction === 'up') {
-    return <ArrowUpRight className="w-4 h-4" />
-  } else if (direction === 'down') {
-    return <ArrowDownRight className="w-4 h-4" />
-  } else {
-    return <Minus className="w-4 h-4" />
-  }
-}
-
-function formatTooltipName(name: string | undefined): string {
-  if (name === 'income') return 'Income'
-  if (name === 'incomeAvg') return 'Income (3m avg)'
-  if (name === 'expenses') return 'Spending'
-  if (name === 'expensesAvg') return 'Spending (3m avg)'
-  if (name === 'savings') return 'Savings'
-  if (name === 'savingsAvg') return 'Savings (3m avg)'
-  return name || ''
-}
-
-// ─── Extracted sub-components ────────────────────────────────────────
-
-interface TrendCardProps {
-  metrics: TrendMetrics
-  icon: React.ElementType
-  iconBgClass: string
-  iconColorClass: string
-  label: string
-  isPositiveGood: boolean
-  delay: number
-  isLoading: boolean
-  valueClassName?: string
-  averageClassName?: string
-  secondStatLabel?: string
-  secondStatClassName?: string
-}
-
-function TrendCard({
-  metrics,
-  icon: Icon,
-  iconBgClass,
-  iconColorClass,
-  label,
-  isPositiveGood,
-  delay,
-  isLoading,
-  valueClassName = 'text-white',
-  averageClassName = 'text-foreground',
-  secondStatLabel = 'Peak',
-  secondStatClassName = 'text-foreground',
-}: Readonly<TrendCardProps>) {
-  const getTrendIcon = (direction: TrendDirection, positiveGood: boolean) => {
-    if (direction === 'stable') return <Minus className="w-5 h-5 text-muted-foreground" />
-    if (direction === 'up') {
-      return positiveGood
-        ? <TrendingUp className="w-5 h-5 text-app-green" />
-        : <TrendingUp className="w-5 h-5 text-app-red" />
-    }
-    return positiveGood
-      ? <TrendingDown className="w-5 h-5 text-app-red" />
-      : <TrendingDown className="w-5 h-5 text-app-green" />
-  }
-
-  const getTrendColor = (direction: TrendDirection, positiveGood: boolean) => {
-    if (direction === 'stable') return 'text-muted-foreground'
-    if (direction === 'up') return positiveGood ? 'text-app-green' : 'text-app-red'
-    return positiveGood ? 'text-app-red' : 'text-app-green'
-  }
-
-  const secondStatValue = metrics.highest
-
-  return (
-    <motion.div
-      initial={{ opacity: 0, y: 20 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ delay }}
-      className="glass rounded-2xl border border-border p-4 md:p-6"
-    >
-      <div className="flex items-center justify-between mb-4">
-        <div className="flex items-center gap-3">
-          <div className={`p-3 ${iconBgClass} rounded-xl`}>
-            <Icon className={`w-6 h-6 ${iconColorClass}`} />
-          </div>
-          <div>
-            <p className="text-sm text-muted-foreground">{label}</p>
-            <p className={`text-xl sm:text-2xl font-bold ${valueClassName}`}>
-              {isLoading ? '...' : formatCurrency(metrics.current)}
-            </p>
-          </div>
-        </div>
-        {!isLoading && getTrendIcon(metrics.direction, isPositiveGood)}
-      </div>
-
-      {!isLoading && (
-        <div className="space-y-3">
-          <div className={`flex items-center gap-2 ${getTrendColor(metrics.direction, isPositiveGood)}`}>
-            {getDirectionIcon(metrics.direction)}
-            <span className="font-semibold">{formatPercent(metrics.changePercent)}</span>
-            <span className="text-text-tertiary text-sm">vs previous month</span>
-          </div>
-          <div className="grid grid-cols-2 gap-2 pt-2 border-t border-border">
-            <div>
-              <p className="text-xs text-text-tertiary">Average</p>
-              <p className={`text-sm font-medium ${averageClassName}`}>{formatCurrency(metrics.average)}</p>
-            </div>
-            <div>
-              <p className="text-xs text-text-tertiary">{secondStatLabel}</p>
-              <p className={`text-sm font-medium ${secondStatClassName}`}>{formatCurrency(secondStatValue)}</p>
-            </div>
-          </div>
-        </div>
-      )}
-    </motion.div>
-  )
-}
-
-interface MonthlyBreakdownTableProps {
-  isLoading: boolean
-  chartData: Array<{ month: string; income: number; expenses: number; surplus: number; rawSavingsRate: number }>
-  sortedChartData: Array<{ month: string; income: number; expenses: number; surplus: number; rawSavingsRate: number }>
-  trendSortKey: string | null
-  trendSortDir: 'asc' | 'desc'
-  toggleTrendSort: (key: string) => void
-}
-
-function MonthlyBreakdownTable({
-  isLoading,
-  chartData,
-  sortedChartData,
-  trendSortKey,
-  trendSortDir,
-  toggleTrendSort,
-}: Readonly<MonthlyBreakdownTableProps>) {
-  return (
-    <motion.div
-      initial={{ opacity: 0, y: 40 }}
-      whileInView={{ opacity: 1, y: 0 }}
-      viewport={{ once: true, margin: '-50px' }}
-      transition={{ duration: 0.6, ease: [0.25, 0.46, 0.45, 0.94] }}
-      className="glass rounded-2xl border border-border p-4 md:p-6"
-    >
-      <h3 className="text-lg font-semibold text-white mb-4">Month-on-Month Breakdown</h3>
-      {isLoading && (
-        <div className="text-center py-8 text-muted-foreground">Loading data...</div>
-      )}
-      {!isLoading && chartData.length > 0 && (
-        <div className="overflow-x-auto">
-          <table className="w-full">
-            <thead>
-              <tr className="border-b border-border">
-                <th className="text-left py-3 px-4 text-sm font-semibold text-muted-foreground">Month</th>
-                <th
-                  onClick={() => toggleTrendSort('income')}
-                  onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); toggleTrendSort('income') } }}
-                  tabIndex={0}
-                  aria-sort={ariaSort(trendSortKey, 'income', trendSortDir)}
-                  className="text-right py-3 px-4 text-sm font-semibold text-muted-foreground cursor-pointer hover:text-white select-none"
-                >
-                  Income {trendSortKey === 'income' && (trendSortDir === 'asc' ? '\u2191' : '\u2193')}
-                </th>
-                <th
-                  onClick={() => toggleTrendSort('expenses')}
-                  onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); toggleTrendSort('expenses') } }}
-                  tabIndex={0}
-                  aria-sort={ariaSort(trendSortKey, 'expenses', trendSortDir)}
-                  className="text-right py-3 px-4 text-sm font-semibold text-muted-foreground cursor-pointer hover:text-white select-none"
-                >
-                  Spending {trendSortKey === 'expenses' && (trendSortDir === 'asc' ? '\u2191' : '\u2193')}
-                </th>
-                <th
-                  onClick={() => toggleTrendSort('surplus')}
-                  onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); toggleTrendSort('surplus') } }}
-                  tabIndex={0}
-                  aria-sort={ariaSort(trendSortKey, 'surplus', trendSortDir)}
-                  className="text-right py-3 px-4 text-sm font-semibold text-muted-foreground cursor-pointer hover:text-white select-none"
-                >
-                  Savings {trendSortKey === 'surplus' && (trendSortDir === 'asc' ? '\u2191' : '\u2193')}
-                </th>
-                <th
-                  onClick={() => toggleTrendSort('rawSavingsRate')}
-                  onKeyDown={(e) => { if (e.key === 'Enter' || e.key === ' ') { e.preventDefault(); toggleTrendSort('rawSavingsRate') } }}
-                  tabIndex={0}
-                  aria-sort={ariaSort(trendSortKey, 'rawSavingsRate', trendSortDir)}
-                  className="text-right py-3 px-4 text-sm font-semibold text-muted-foreground cursor-pointer hover:text-white select-none"
-                >
-                  Savings Rate {trendSortKey === 'rawSavingsRate' && (trendSortDir === 'asc' ? '\u2191' : '\u2193')}
-                </th>
-              </tr>
-            </thead>
-            <motion.tbody
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              transition={{ duration: 0.2 }}
-            >
-              {sortedChartData.map((trend) => (
-                <tr
-                  key={trend.month}
-                  className="border-b border-border hover:bg-white/10 transition-colors"
-                >
-                  <td className="py-3 px-4 text-white font-medium">{trend.month}</td>
-                  <td className="py-3 px-4 text-right text-app-green">{formatCurrency(trend.income)}</td>
-                  <td className="py-3 px-4 text-right text-app-red">{formatCurrency(trend.expenses)}</td>
-                  <td className={`py-3 px-4 text-right font-bold ${trend.surplus >= 0 ? 'text-app-purple' : 'text-app-red'}`}>
-                    {formatCurrency(trend.surplus)}
-                  </td>
-                  <td className={`py-3 px-4 text-right ${trend.rawSavingsRate >= 0 ? 'text-foreground' : 'text-app-red'}`}>
-                    {trend.rawSavingsRate.toFixed(1)}%
-                  </td>
-                </tr>
-              ))}
-            </motion.tbody>
-          </table>
-        </div>
-      )}
-      {!isLoading && chartData.length === 0 && (
-        <EmptyState
-          icon={TrendingUp}
-          title="No data available"
-          description="Monthly breakdown will appear here once you have transactions."
-          variant="compact"
-        />
-      )}
-    </motion.div>
-  )
-}
-
-function ariaSort(activeKey: string | null, column: string, dir: 'asc' | 'desc'): 'ascending' | 'descending' | 'none' {
-  if (activeKey !== column) return 'none'
-  return dir === 'asc' ? 'ascending' : 'descending'
-}
-
-export default function TrendsForecastsPage() {
-  const dims = useChartDimensions()
-  const { data: preferences } = usePreferences()
-  const savingsGoalPercent = preferences?.savings_goal_percent ?? 20
-
-  // Fetch all trends data; filter client-side by date range
-  const { data: trendsData, isLoading } = useTrends('all_time')
-  const { data: allTransactions = [] } = useTransactions()
-
-  const { dateRange, timeFilterProps } = useAnalyticsTimeFilter(
-    allTransactions,
-    { availableModes: ['all_time', 'fy', 'yearly'] },
-  )
-
-  // Filter monthly trends by the selected date range
-  const filteredMonthlyTrends = useMemo(() => {
-    if (!trendsData?.monthly_trends) return []
-    if (!dateRange.start_date) return trendsData.monthly_trends
-
-    return trendsData.monthly_trends.filter((t) => {
-      // monthly_trends have a 'month' field like "2024-01" — compare as YYYY-MM
-      const monthStart = `${t.month}-01`
-      if (dateRange.start_date && monthStart < dateRange.start_date.substring(0, 10)) return false
-      if (dateRange.end_date && monthStart > dateRange.end_date.substring(0, 10)) return false
-      return true
-    })
-  }, [trendsData, dateRange])
-
-  // Calculate comprehensive trend metrics
-  const metrics = useMemo(() => {
-    const defaultMetrics: TrendMetrics = {
-      current: 0,
-      previous: 0,
-      change: 0,
-      changePercent: 0,
-      direction: 'stable',
-      average: 0,
-      highest: 0,
-      lowest: 0,
-    }
-
-    if (!filteredMonthlyTrends || filteredMonthlyTrends.length < 1) {
-      return {
-        spending: defaultMetrics,
-        income: defaultMetrics,
-        savings: defaultMetrics,
-      }
-    }
-
-    const trends = filteredMonthlyTrends
-    const latest = trends.at(-1)
-    if (!latest) return { spending: defaultMetrics, income: defaultMetrics, savings: defaultMetrics }
-    const previous = trends.length > 1 ? (trends.at(-2) ?? latest) : latest
-
-    // Calculate spending metrics
-    const expenses = trends.map(t => t.expenses)
-    const spendingChange = latest.expenses - previous.expenses
-    const spendingChangePercent = percentChange(latest.expenses, previous.expenses) ?? 0
-
-    // Calculate income metrics
-    const incomes = trends.map(t => t.income)
-    const incomeChange = latest.income - previous.income
-    const incomeChangePercent = percentChange(latest.income, previous.income) ?? 0
-
-    // Calculate savings metrics
-    const surpluses = trends.map(t => t.surplus)
-    const savingsChange = latest.surplus - previous.surplus
-    const savingsChangePercent = percentChange(latest.surplus, previous.surplus) ?? 0
-
-    const getDirection = (change: number): TrendDirection => {
-      if (Math.abs(change) < 2) return 'stable'
-      return change > 0 ? 'up' : 'down'
-    }
-
-    return {
-      spending: {
-        current: latest.expenses,
-        previous: previous.expenses,
-        change: spendingChange,
-        changePercent: spendingChangePercent,
-        direction: getDirection(spendingChangePercent),
-        average: expenses.reduce((a, b) => a + b, 0) / expenses.length,
-        highest: Math.max(...expenses),
-        lowest: Math.min(...expenses),
-      },
-      income: {
-        current: latest.income,
-        previous: previous.income,
-        change: incomeChange,
-        changePercent: incomeChangePercent,
-        direction: getDirection(incomeChangePercent),
-        average: incomes.reduce((a, b) => a + b, 0) / incomes.length,
-        highest: Math.max(...incomes),
-        lowest: Math.min(...incomes),
-      },
-      savings: {
-        current: latest.surplus,
-        previous: previous.surplus,
-        change: savingsChange,
-        changePercent: savingsChangePercent,
-        direction: getDirection(savingsChangePercent),
-        average: surpluses.reduce((a, b) => a + b, 0) / surpluses.length,
-        highest: Math.max(...surpluses),
-        lowest: Math.min(...surpluses),
-      },
-    }
-  }, [filteredMonthlyTrends])
-
-  // Prepare chart data for individual trend lines
-  const chartData = useMemo(() => {
-    if (!filteredMonthlyTrends.length) return []
-
-    const rawData = filteredMonthlyTrends.map((t, index, arr) => {
-      const prev = index > 0 ? arr[index - 1] : t
-      const rawSavingsRate = t.income > 0 ? (t.surplus / t.income) * 100 : 0
-      return {
-        ...t,
-        spendingChange: index > 0 ? (percentChange(t.expenses, prev.expenses) ?? 0) : 0,
-        incomeChange: index > 0 ? (percentChange(t.income, prev.income) ?? 0) : 0,
-        rawSavingsRate,
-        // Clamp negative savings rate to 0 for cleaner visualization
-        savingsRate: Math.max(0, rawSavingsRate),
-      }
-    })
-
-    return rawData
-  }, [filteredMonthlyTrends])
-
-  // Filter transactions by the selected date range
-  const filteredTransactions = useMemo(() => {
-    if (!allTransactions.length) return []
-    const startDate = dateRange.start_date
-    if (!startDate) return allTransactions
-
-    return allTransactions.filter((t) => {
-      const txDate = getDateKey(t.date)
-      return txDate >= startDate && (!dateRange.end_date || txDate <= dateRange.end_date)
-    })
-  }, [allTransactions, dateRange])
-
-  // Daily cumulative savings rate data
-  const dailySavingsData = useMemo(() => {
-    if (!filteredTransactions.length) return []
-
-    const dailyMap: Record<string, { income: number; expense: number }> = {}
-    for (const tx of filteredTransactions) {
-      const day = tx.date.substring(0, 10)
-      if (!dailyMap[day]) dailyMap[day] = { income: 0, expense: 0 }
-      if (tx.type === 'Income') dailyMap[day].income += tx.amount
-      else if (tx.type === 'Expense') dailyMap[day].expense += tx.amount
-    }
-
-    const sortedDays = Object.entries(dailyMap).sort(([a], [b]) => a.localeCompare(b))
-    let cumIncome = 0
-    let cumExpense = 0
-
-    return sortedDays.map(([date, { income, expense }]) => {
-      cumIncome += income
-      cumExpense += expense
-      const savingsRate = cumIncome > 0 ? ((cumIncome - cumExpense) / cumIncome) * 100 : 0
-      return {
-        date,
-        savingsRate: Math.max(0, savingsRate),
-        rawSavingsRate: savingsRate,
-      }
-    })
-  }, [filteredTransactions])
-
-  // Monthly trend data for the small multiples charts
-  const monthlyTrendChartData = useMemo(() => {
-    if (!filteredMonthlyTrends.length) return []
-    return filteredMonthlyTrends.map((t) => ({
-      month: t.month,
-      label: new Date(t.month + '-01').toLocaleDateString('en-US', { month: 'short', year: '2-digit' }),
-      income: t.income,
-      expenses: t.expenses,
-      savings: t.surplus,
-    }))
-  }, [filteredMonthlyTrends])
-
-  // 3-month rolling average for the monthly charts
-  const monthlyTrendWithAvg = useMemo(() => {
-    return monthlyTrendChartData.map((d, i) => {
-      const start = Math.max(0, i - 2)
-      const window = monthlyTrendChartData.slice(start, i + 1)
-      return {
-        ...d,
-        incomeAvg: window.reduce((s, w) => s + w.income, 0) / window.length,
-        expensesAvg: window.reduce((s, w) => s + w.expenses, 0) / window.length,
-        savingsAvg: window.reduce((s, w) => s + w.savings, 0) / window.length,
-      }
-    })
-  }, [monthlyTrendChartData])
-
-  // Peak values for reference lines (monthly)
-  const peakIncome = useMemo(() => Math.max(...monthlyTrendChartData.map(d => d.income), 0), [monthlyTrendChartData])
-  const peakExpenses = useMemo(() => Math.max(...monthlyTrendChartData.map(d => d.expenses), 0), [monthlyTrendChartData])
-  const peakSavings = useMemo(() => Math.max(...monthlyTrendChartData.map(d => d.savings), 0), [monthlyTrendChartData])
-
-  // Sorting state for the monthly breakdown table
-  const [trendSortKey, setTrendSortKey] = useState<string | null>(null)
-  const [trendSortDir, setTrendSortDir] = useState<'asc' | 'desc'>('desc')
-
-  const toggleTrendSort = (key: string) => {
-    if (trendSortKey === key) {
-      setTrendSortDir(trendSortDir === 'asc' ? 'desc' : 'asc')
-    } else {
-      setTrendSortKey(key)
-      setTrendSortDir('desc')
-    }
-  }
-
-  const sortedChartData = useMemo(() => {
-    const data = chartData.slice(-8)
-    if (!trendSortKey) return data
-    return [...data].sort((a, b) => {
-      const av = a[trendSortKey as keyof typeof a]
-      const bv = b[trendSortKey as keyof typeof b]
-      const cmp = typeof av === 'number' && typeof bv === 'number' ? av - bv : String(av).localeCompare(String(bv))
-      return trendSortDir === 'asc' ? cmp : -cmp
-    })
-  }, [chartData, trendSortKey, trendSortDir])
-
-  // Linked crosshair state for small-multiples charts
-  const [activeLabel, setActiveLabel] = useState<string | null>(null)
-
-  return (
-    <div className="min-h-screen p-4 md:p-6 lg:p-8">
-      <div className="max-w-7xl mx-auto space-y-6 md:space-y-8">
-        <PageHeader
-          title="Trends & Forecasts"
-          subtitle="Analyze patterns and predict future trends"
-          action={
-            <AnalyticsTimeFilter {...timeFilterProps} />
-          }
-        />
-
-        {/* Trend Cards */}
-        <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4 sm:gap-6">
-          <TrendCard
-            metrics={metrics.spending}
-            icon={CreditCard}
-            iconBgClass="bg-app-red/20"
-            iconColorClass="text-app-red"
-            label="Spending Trend"
-            isPositiveGood={false}
-            delay={0.2}
-            isLoading={isLoading}
-          />
-          <TrendCard
-            metrics={metrics.income}
-            icon={Wallet}
-            iconBgClass="bg-app-green/20"
-            iconColorClass="text-app-green"
-            label="Income Trend"
-            isPositiveGood={true}
-            delay={0.3}
-            isLoading={isLoading}
-          />
-          <TrendCard
-            metrics={metrics.savings}
-            icon={PiggyBank}
-            iconBgClass="bg-app-purple/20"
-            iconColorClass="text-app-purple"
-            label="Savings Trend"
-            isPositiveGood={true}
-            delay={0.4}
-            isLoading={isLoading}
-            valueClassName={metrics.savings.current >= 0 ? 'text-white' : 'text-app-red'}
-            averageClassName={metrics.savings.average >= 0 ? 'text-foreground' : 'text-app-red'}
-            secondStatLabel="Best Month"
-            secondStatClassName="text-app-green"
-          />
-        </div>
-
-        {/* Main Trend Chart — Small Multiples */}
-        <motion.div
-          initial={{ opacity: 0, y: 40 }}
-          whileInView={{ opacity: 1, y: 0 }}
-          viewport={{ once: true, margin: '-50px' }}
-          transition={{ duration: 0.6, ease: [0.25, 0.46, 0.45, 0.94] }}
-          className="glass rounded-2xl border border-border p-4 md:p-6"
-        >
-          <div className="flex items-center gap-3 mb-6">
-            <LineChart className="w-5 h-5 text-app-blue" />
-            <div>
-              <h3 className="text-lg font-semibold text-white">Income & Expense Trends</h3>
-              <p className="text-sm text-text-tertiary">Monthly breakdown with 3-month rolling averages</p>
-            </div>
-          </div>
-          {isLoading && (
-            <div className="h-80 flex items-center justify-center">
-              <div className="animate-pulse text-muted-foreground">Loading chart...</div>
-            </div>
-          )}
-          {!isLoading && monthlyTrendWithAvg.length > 0 && (
-            <div className="grid grid-cols-1 lg:grid-cols-3 gap-4">
-              {/* Income mini chart */}
-              <div className="glass-thin rounded-xl border border-border p-4">
-                <div className="flex items-center gap-2 mb-3">
-                  <div className="w-3 h-3 rounded-full bg-app-green" />
-                  <span className="text-sm font-medium text-white">Income</span>
-                </div>
-                {monthlyTrendWithAvg.length === 0 ? (
-                  <ChartEmptyState height={180} />
-                ) : (
-                  <ChartContainer height={180}>
-                    <AreaChart data={monthlyTrendWithAvg} onMouseMove={(e) => { if (e?.activeLabel) setActiveLabel(e.activeLabel as string) }} onMouseLeave={() => setActiveLabel(null)}>
-                      <defs>
-                        {areaGradient('trendIncome', rawColors.app.green, 0.4, 0.02)}
-                      </defs>
-                      <CartesianGrid {...GRID_DEFAULTS} />
-                      <XAxis {...xAxisDefaults(monthlyTrendWithAvg.length)} dataKey="label" />
-                      <YAxis hide />
-                      <Tooltip
-                        {...chartTooltipProps}
-                        labelFormatter={(_label: unknown, payload: ReadonlyArray<{ payload?: { month?: string } }>) => {
-                          const month = payload?.[0]?.payload?.month
-                          return month ? new Date(month + '-01').toLocaleDateString('en-US', { month: 'long', year: 'numeric' }) : ''
-                        }}
-                        formatter={(value: number | undefined, name: string | undefined) => [
-                          value === undefined ? '' : formatCurrency(value),
-                          formatTooltipName(name)
-                        ]}
-                      />
-                      <ReferenceLine y={peakIncome} stroke="rgba(255,255,255,0.2)" strokeDasharray="3 3" label={{ value: `Peak: ${formatCurrencyShort(peakIncome)}`, fill: '#71717a', fontSize: 10, position: 'insideTopRight' }} />
-                      {activeLabel && <ReferenceLine x={activeLabel} stroke="rgba(255,255,255,0.3)" strokeDasharray="3 3" />}
-                      <Area type="monotone" dataKey="income" stroke={rawColors.app.green} fill={areaGradientUrl('trendIncome')} strokeWidth={2} dot={false} activeDot={{ ...ACTIVE_DOT, fill: rawColors.app.green }} isAnimationActive={shouldAnimate(monthlyTrendWithAvg.length)} animationDuration={600} animationEasing="ease-out" />
-                      <Line type="monotone" dataKey="incomeAvg" stroke={rawColors.app.green} strokeWidth={2} strokeDasharray="6 3" dot={false} activeDot={{ ...ACTIVE_DOT, fill: rawColors.app.green }} name="Income (3m avg)" isAnimationActive={shouldAnimate(monthlyTrendWithAvg.length)} animationDuration={600} animationEasing="ease-out" />
-                    </AreaChart>
-                  </ChartContainer>
-                )}
-              </div>
-
-              {/* Expenses mini chart */}
-              <div className="glass-thin rounded-xl border border-border p-4">
-                <div className="flex items-center gap-2 mb-3">
-                  <div className="w-3 h-3 rounded-full bg-app-red" />
-                  <span className="text-sm font-medium text-white">Expenses</span>
-                </div>
-                {monthlyTrendWithAvg.length === 0 ? (
-                  <ChartEmptyState height={180} />
-                ) : (
-                  <ChartContainer height={180}>
-                    <AreaChart data={monthlyTrendWithAvg} onMouseMove={(e) => { if (e?.activeLabel) setActiveLabel(e.activeLabel as string) }} onMouseLeave={() => setActiveLabel(null)}>
-                      <defs>
-                        {areaGradient('trendExpense', rawColors.app.red, 0.4, 0.02)}
-                      </defs>
-                      <CartesianGrid {...GRID_DEFAULTS} />
-                      <XAxis {...xAxisDefaults(monthlyTrendWithAvg.length)} dataKey="label" />
-                      <YAxis hide />
-                      <Tooltip
-                        {...chartTooltipProps}
-                        labelFormatter={(_label: unknown, payload: ReadonlyArray<{ payload?: { month?: string } }>) => {
-                          const month = payload?.[0]?.payload?.month
-                          return month ? new Date(month + '-01').toLocaleDateString('en-US', { month: 'long', year: 'numeric' }) : ''
-                        }}
-                        formatter={(value: number | undefined, name: string | undefined) => [
-                          value === undefined ? '' : formatCurrency(value),
-                          formatTooltipName(name)
-                        ]}
-                      />
-                      <ReferenceLine y={peakExpenses} stroke="rgba(255,255,255,0.2)" strokeDasharray="3 3" label={{ value: `Peak: ${formatCurrencyShort(peakExpenses)}`, fill: '#71717a', fontSize: 10, position: 'insideTopRight' }} />
-                      {activeLabel && <ReferenceLine x={activeLabel} stroke="rgba(255,255,255,0.3)" strokeDasharray="3 3" />}
-                      <Area type="monotone" dataKey="expenses" stroke={rawColors.app.red} fill={areaGradientUrl('trendExpense')} strokeWidth={2} dot={false} activeDot={{ ...ACTIVE_DOT, fill: rawColors.app.red }} isAnimationActive={shouldAnimate(monthlyTrendWithAvg.length)} animationDuration={600} animationEasing="ease-out" />
-                      <Line type="monotone" dataKey="expensesAvg" stroke={rawColors.app.red} strokeWidth={2} strokeDasharray="6 3" dot={false} activeDot={{ ...ACTIVE_DOT, fill: rawColors.app.red }} name="Spending (3m avg)" isAnimationActive={shouldAnimate(monthlyTrendWithAvg.length)} animationDuration={600} animationEasing="ease-out" />
-                    </AreaChart>
-                  </ChartContainer>
-                )}
-              </div>
-
-              {/* Savings mini chart */}
-              <div className="glass-thin rounded-xl border border-border p-4">
-                <div className="flex items-center gap-2 mb-3">
-                  <div className="w-3 h-3 rounded-full bg-app-purple" />
-                  <span className="text-sm font-medium text-white">Savings</span>
-                </div>
-                {monthlyTrendWithAvg.length === 0 ? (
-                  <ChartEmptyState height={180} />
-                ) : (
-                  <ChartContainer height={180}>
-                    <AreaChart data={monthlyTrendWithAvg} onMouseMove={(e) => { if (e?.activeLabel) setActiveLabel(e.activeLabel as string) }} onMouseLeave={() => setActiveLabel(null)}>
-                      <defs>
-                        {areaGradient('trendSavings', rawColors.app.purple, 0.4, 0.02)}
-                      </defs>
-                      <CartesianGrid {...GRID_DEFAULTS} />
-                      <XAxis {...xAxisDefaults(monthlyTrendWithAvg.length)} dataKey="label" />
-                      <YAxis hide />
-                      <Tooltip
-                        {...chartTooltipProps}
-                        labelFormatter={(_label: unknown, payload: ReadonlyArray<{ payload?: { month?: string } }>) => {
-                          const month = payload?.[0]?.payload?.month
-                          return month ? new Date(month + '-01').toLocaleDateString('en-US', { month: 'long', year: 'numeric' }) : ''
-                        }}
-                        formatter={(value: number | undefined, name: string | undefined) => [
-                          value === undefined ? '' : formatCurrency(value),
-                          formatTooltipName(name)
-                        ]}
-                      />
-                      <ReferenceLine y={peakSavings} stroke="rgba(255,255,255,0.2)" strokeDasharray="3 3" label={{ value: `Peak: ${formatCurrencyShort(peakSavings)}`, fill: '#71717a', fontSize: 10, position: 'insideTopRight' }} />
-                      {activeLabel && <ReferenceLine x={activeLabel} stroke="rgba(255,255,255,0.3)" strokeDasharray="3 3" />}
-                      <Area type="monotone" dataKey="savings" stroke={rawColors.app.purple} fill={areaGradientUrl('trendSavings')} strokeWidth={2} dot={false} activeDot={{ ...ACTIVE_DOT, fill: rawColors.app.purple }} isAnimationActive={shouldAnimate(monthlyTrendWithAvg.length)} animationDuration={600} animationEasing="ease-out" />
-                      <Line type="monotone" dataKey="savingsAvg" stroke={rawColors.app.purple} strokeWidth={2} strokeDasharray="6 3" dot={false} activeDot={{ ...ACTIVE_DOT, fill: rawColors.app.purple }} name="Savings (3m avg)" isAnimationActive={shouldAnimate(monthlyTrendWithAvg.length)} animationDuration={600} animationEasing="ease-out" />
-                    </AreaChart>
-                  </ChartContainer>
-                )}
-              </div>
-            </div>
-          )}
-          {!isLoading && monthlyTrendWithAvg.length === 0 && (
-            <EmptyState
-              icon={LineChart}
-              title="No data available"
-              description="Upload your transaction data to see spending trends and forecasts."
-              actionLabel="Upload Data"
-              actionHref="/upload"
-              variant="chart"
-            />
-          )}
-        </motion.div>
-
-        {/* Savings Rate Trend */}
-        <motion.div
-          initial={{ opacity: 0, y: 40 }}
-          whileInView={{ opacity: 1, y: 0 }}
-          viewport={{ once: true, margin: '-50px' }}
-          transition={{ duration: 0.6, ease: [0.25, 0.46, 0.45, 0.94] }}
-          className="glass rounded-2xl border border-border p-4 md:p-6"
-        >
-          <div className="flex items-center gap-3 mb-6">
-            <PiggyBank className="w-5 h-5 text-app-purple" />
-            <h3 className="text-lg font-semibold text-white">Savings Rate Trend</h3>
-            <span className="text-sm text-text-tertiary">(% of income saved each month)</span>
-          </div>
-          {isLoading && (
-            <div className="h-64 flex items-center justify-center">
-              <div className="animate-pulse text-muted-foreground">Loading chart...</div>
-            </div>
-          )}
-          {!isLoading && dailySavingsData.length > 0 && (
-            dailySavingsData.length === 0 ? (
-              <ChartEmptyState height={250} />
-            ) : (
-              <ChartContainer height={250}>
-                <AreaChart data={dailySavingsData}>
-                  <defs>
-                    {areaGradient('savingsRate', rawColors.app.purple, 0.4, 0.02)}
-                  </defs>
-                  <CartesianGrid {...GRID_DEFAULTS} />
-                  <XAxis {...xAxisDefaults(dailySavingsData.length, { angle: dims.angleXLabels ? -45 : undefined, height: 70, dateFormatter: true })} dataKey="date" />
-                  <YAxis {...yAxisDefaults({ currency: false })} tickFormatter={(v: number) => `${Math.round(v)}%`} domain={[0, 'auto']} />
-                  <Tooltip
-                    {...chartTooltipProps}
-                    labelFormatter={(label) => new Date(label).toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })}
-                    formatter={(_value: number | undefined, _name: string | undefined, props: { payload?: { rawSavingsRate?: number } }) => {
-                      const actual = props.payload?.rawSavingsRate ?? 0
-                      const label = actual < 0 ? `${actual.toFixed(1)}% (deficit)` : `${actual.toFixed(1)}%`
-                      return [label, 'Cumulative Savings Rate']
-                    }}
-                  />
-                  <ReferenceLine
-                    y={savingsGoalPercent}
-                    stroke={rawColors.app.green}
-                    strokeDasharray="6 4"
-                    strokeWidth={1.5}
-                    label={{
-                      value: `Target: ${savingsGoalPercent}%`,
-                      fill: rawColors.app.green,
-                      fontSize: 11,
-                      position: 'insideTopRight',
-                    }}
-                  />
-                  <Area
-                    type="monotone"
-                    dataKey="savingsRate"
-                    stroke={rawColors.app.purple}
-                    fill={areaGradientUrl('savingsRate')}
-                    strokeWidth={2}
-                    dot={false}
-                    isAnimationActive={shouldAnimate(dailySavingsData.length)}
-                    animationDuration={600}
-                    animationEasing="ease-out"
-                  />
-                </AreaChart>
-              </ChartContainer>
-            )
-          )}
-          {!isLoading && dailySavingsData.length === 0 && (
-            <EmptyState
-              icon={PiggyBank}
-              title="No data available"
-              description="Add transactions to track your savings rate over time."
-              variant="compact"
-            />
-          )}
-        </motion.div>
-
-        {/* Monthly Breakdown Table */}
-        <MonthlyBreakdownTable
-          isLoading={isLoading}
-          chartData={chartData}
-          sortedChartData={sortedChartData}
-          trendSortKey={trendSortKey}
-          trendSortDir={trendSortDir}
-          toggleTrendSort={toggleTrendSort}
-        />
-
-        {/* Cash Flow Forecast */}
-        <CashFlowForecast />
-      </div>
-    </div>
-  )
-}
