+++ frontend/src/pages/trends-forecasts/types.ts
@@ -1,0 +1,20 @@
+export type TrendDirection = 'up' | 'down' | 'stable'
+
+export interface TrendMetrics {
+  current: number
+  previous: number
+  change: number
+  changePercent: number
+  direction: TrendDirection
+  average: number
+  highest: number
+  lowest: number
+}
+
+export interface MonthlyTrendRow {
+  month: string
+  income: number
+  expenses: number
+  surplus: number
+  rawSavingsRate: number
+}
